+++ mvs_eland_tool/mvs_eland_tool.py
@@ -62,7 +62,9 @@ def main(**kwargs):
     # Parse the arguments from the command line
     parser = initializing.create_parser()
     args = vars(parser.parse_args())
-    kwargs.update(**args)
+    # Give priority from kwargs on command line arguments
+    args.update(**kwargs)
+    kwargs = args
 
     logging.debug("Accessing script: A0_initialization")
     user_input = initializing.welcome(welcome_text, **kwargs)
@@ -74,7 +76,9 @@ def main(**kwargs):
 
     if not user_input["path_input_file"].endswith("json"):
         logging.debug("Accessing script: A1_csv_to_json")
-        path_to_json_from_csv = load_data_from_csv.create_input_json()
+        path_to_json_from_csv = load_data_from_csv.create_input_json(
+            input_directory=user_input["path_input_folder"]
+        )
         user_input.update({"path_input_file": path_to_json_from_csv})
 
     logging.debug("Accessing script: B0_data_input_json")