+++ CHANGELOG.md
@@ -21,8 +21,12 @@ Here is a template for new release sections
 ### Added
 
 ### Changed
+- Give priority from kwargs on command line arguments (#112)
 
 ### Removed
+
+### Fixed
+- Input directory of csv files specified by user is handed to `load_data_from_csv.create_input_json()` (#112)
 
 ## [0.1.0] -2020-01-30